fix: register akmods key after it has been created
There script was trying to register the akmods key to the system
before akmods had run for the first time and created its key,
so there was no key to register.
Also, this was not needed since it hadn't signed any module yet.

The script has been moved after the installations of the nVidia driver,
which is signed by akmods after installing.